wait pid fixed ?

+++ Userland/SampleCodeModule/user_syscalls.c
@@ -92,7 +92,7 @@ void usys_yield(){
 }
 
 void usys_wait_processes(int16_t pid){
-    _wait_processes(pid);
+    _wait_processes((uint64_t)pid);
 }
 
 void usys_ps(){
